# [nnx] optimize NodeDef.attributes