--- a/flax/nnx/scripts/run-all-examples.bash
+++ b/flax/nnx/scripts/run-all-examples.bash
@@ -1,9 +1,8 @@
 set -e
 
 source .venv/bin/activate
-cd flax/nnx
 
-for f in $(find examples/toy_examples -name "*.py" -maxdepth 1); do
+for f in $(find examples/nnx_toy_examples -name "*.py" -maxdepth 1); do
     echo -e "\n---------------------------------"
     echo "$f"
     echo "---------------------------------"