--- a/tests/nnx/graph_utils_test.py
+++ b/tests/nnx/graph_utils_test.py
@@ -303,7 +303,7 @@ class TestGraphUtils(absltest.TestCase):
 
     assert 'tree' in state
     assert 'a' in state.tree
-    assert graphdef.subgraphs['tree'].type is nnx.graph.GenericPytree
+    assert graphdef.attributes[0].value.type is nnx.graph.GenericPytree
 
     m2 = nnx.merge(graphdef, state)