… docs_nnx/nnx_basics.ipynb