# What does this PR do?

```
python benchmarks/nnx_graph_overhead.py --mode=all --depth=100 --total_steps=1000

### NNX ###
total time: 7.961960315704346
time per step: 7961.96 µs
time per layer: 79.62 µs

### JAX ###
total time: 1.789884328842163
time per step: 1789.88 µs
time per layer: 17.90 µs
```